requirements.txt
# Web Scraping e Automação
selenium==4.16.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
lxml==5.1.0

# Google Cloud Platform
google-cloud-storage==2.14.0
google-cloud-bigquery==3.14.1
google-cloud-secret-manager==2.18.0

# Processamento de Dados
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2
xlrd==2.0.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Logging e Monitoramento
loguru==0.7.2

# Validação e Configuração
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0






